fix remaining clippy issues with a small reorg
